split special perf
